requirements.txt
# Semantic-Forest-Lab dependency set (Python 3.8 compatible)
#
# NOTE:
# - PyTorch is optional and should be installed separately per CUDA/CPU environment.
# - For conda users, RDKit is recommended from conda-forge.

# Core numeric stack (pin to avoid C-API / ABI mismatch)
numpy==1.24.4
pandas==2.0.3
numexpr==2.8.6
scikit-learn==1.3.2

# Ontology / chemistry
owlready2>=0.46,<0.47
rdflib>=6.2,<7.0
rdkit-pypi>=2022.9.5
PyYAML>=6.0,<7.0

# Visualization
matplotlib>=3.6,<3.8
seaborn>=0.12,<0.14

# Hyperparameter optimization
optuna>=3.6,<5.0
